+++ app/src/main/java/woowacourse/payments/ui/component/CardNumberInputField.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.registercard.component
+package woowacourse.payments.ui.component
 
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.text.KeyboardOptions
