+++ app/src/main/java/woowacourse/payments/ui/component/ExpiryDateInputField.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.registercard.component
+package woowacourse.payments.ui.component
 
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.text.KeyboardOptions
@@ -9,12 +9,12 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import woowacourse.payments.R
+import woowacourse.payments.ui.registercard.ExpiryDateVisualTransformation
 
 @Composable
-fun PasswordInputField(
+fun ExpiryDateInputField(
     text: String,
     onValueChange: (String) -> Unit,
     isError: Boolean,
@@ -23,14 +23,14 @@ fun PasswordInputField(
         modifier = Modifier.fillMaxWidth(0.5f),
         value = text,
         onValueChange = onValueChange,
-        label = { Text(stringResource(R.string.password_label)) },
+        label = { Text(stringResource(R.string.expiry_date_label)) },
         placeholder = {
             Text(
-                text = stringResource(R.string.password_place_holder),
+                text = stringResource(R.string.expiry_date_place_holder),
                 color = Color.Gray,
             )
         },
-        visualTransformation = PasswordVisualTransformation(),
+        visualTransformation = ExpiryDateVisualTransformation(),
         singleLine = true,
         keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
         isError = isError,
@@ -39,8 +39,8 @@ fun PasswordInputField(
 
 @Preview(showBackground = true)
 @Composable
-private fun PasswordInputFieldPreview() {
-    PasswordInputField(
+private fun ExpiryDateInputFieldPreview() {
+    ExpiryDateInputField(
         text = "",
         onValueChange = { },
         isError = false,
